consolidate theme providers
