+++ apps/sim/app/layout.tsx
@@ -9,7 +9,7 @@ import { createLogger } from '@/lib/logs/console/logger'
 import '@/app/globals.css'
 
 import { SessionProvider } from '@/lib/session/session-context'
-import { ConditionalThemeProvider } from '@/app/conditional-theme-provider'
+import { ThemeProvider } from '@/app/theme-provider'
 import { ZoomPrevention } from '@/app/zoom-prevention'
 
 const logger = createLogger('RootLayout')
@@ -91,7 +91,7 @@ export default function RootLayout({ children }: { children: React.ReactNode }) 
         <PublicEnvScript />
       </head>
       <body suppressHydrationWarning>
-        <ConditionalThemeProvider>
+        <ThemeProvider>
           <SessionProvider>
             <BrandedLayout>
               <ZoomPrevention />
@@ -103,7 +103,7 @@ export default function RootLayout({ children }: { children: React.ReactNode }) 
               )}
             </BrandedLayout>
           </SessionProvider>
-        </ConditionalThemeProvider>
+        </ThemeProvider>
       </body>
     </html>
   )
