+++ apps/sim/app/theme-provider.tsx
@@ -1,9 +1,27 @@
 'use client'
 
+import { usePathname } from 'next/navigation'
 import type { ThemeProviderProps } from 'next-themes'
 import { ThemeProvider as NextThemesProvider } from 'next-themes'
 
 export function ThemeProvider({ children, ...props }: ThemeProviderProps) {
+  const pathname = usePathname()
+
+  // Force light mode for certain pages
+  const forcedTheme =
+    pathname === '/' ||
+    pathname === '/homepage' ||
+    pathname.startsWith('/login') ||
+    pathname.startsWith('/signup') ||
+    pathname.startsWith('/terms') ||
+    pathname.startsWith('/privacy') ||
+    pathname.startsWith('/invite') ||
+    pathname.startsWith('/verify') ||
+    pathname.startsWith('/changelog') ||
+    pathname.startsWith('/chat')
+      ? 'light'
+      : undefined
+
   return (
     <NextThemesProvider
       attribute='class'
@@ -11,6 +29,7 @@ export function ThemeProvider({ children, ...props }: ThemeProviderProps) {
       enableSystem
       disableTransitionOnChange
       storageKey='sim-theme'
+      forcedTheme={forcedTheme}
       {...props}
     >
       {children}
